Code execution fix
