+++ clickable-demos/tsconfig.json
@@ -12,6 +12,6 @@
     "types": ["html2canvas", "tippy.js"],
     "skipLibCheck": true
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*.ts", "**/*.d.ts"],
   "exclude": ["node_modules"]
 }
